# contracts/openapi/openapi.yaml
"openapi": "3.1.0"
"info":
  "title": "SmartQuarter max-gateway"
  "version": "1.1.0"
  "description": "Gateway API with real Identity, Issue and Community services. Session actor and active house are server-owned. House registration requires explicit platform administrator approval. Stable HTTP errors map gRPC status codes."
"servers":
  -
    "url": "http://localhost:18080"
"security":
  -
    "SessionAuth": []
"paths":
...
  "/api/v1/house-registrations":
    "post":
      "summary": "CreateHouseRegistration"
      "operationId": "post__api_v1_house-registrations"
      "description": "Required role: Authenticated user, including without membership. gRPC: Identity.HouseService.CreateHouseRegistration.  List responses are bounded to 100 records (members: 500). Mutation requests require a JSON object, including {} for commands without fields."
      "responses":
        "202":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HouseRegistration"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "Origin"
          "in": "header"
          "required": true
          "schema":
            "type": "string"
            "format": "uri"
          "description": "Exact origin from TRUSTED_ORIGINS. Missing or foreign Origin is rejected (CSRF protection)."
        -
          "name": "Idempotency-Key"
          "in": "header"
          "schema":
            "type": "string"
            "format": "uuid"
          "description": "Optional. Scope: user + active house. Key binds method, path and compact JSON bytes (key order significant). Same request replays success for 24h; changed payload 409 IDEMPOTENCY_KEY_REUSED; concurrent/ambiguous outcome 409 IDEMPOTENCY_IN_PROGRESS. Reconcile before creating a new key; no exactly-once guarantee."
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "type": "object"
              "additionalProperties": false
              "properties":
                "name":
                  "type": "string"
                  "maxLength": 255
                "address":
                  "type": "string"
                  "maxLength": 1000
                "city":
                  "type": "string"
                  "maxLength": 100
              "required":
                - "name"
                - "address"
                - "city"
    "get":
      "summary": "ListMyHouseRegistrations"
      "operationId": "get__api_v1_house-registrations"
      "description": "Required role: Authenticated user, including without membership. gRPC: Identity.HouseService.ListMyHouseRegistrations.  List responses are bounded to 100 records (members: 500). Mutation requests require a JSON object, including {} for commands without fields."
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HouseRegistrationList"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
  "/api/v1/house-registrations/{id}":
    "get":
      "summary": "GetHouseRegistration"
      "operationId": "get__api_v1_house-registrations_id"
      "description": "Required role: Authenticated user, including without membership. gRPC: Identity.HouseService.GetHouseRegistration.  List responses are bounded to 100 records (members: 500). Mutation requests require a JSON object, including {} for commands without fields."
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HouseRegistration"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "format": "uuid"
  "/api/v1/house-registrations/{id}/cancel":
    "post":
      "summary": "CancelHouseRegistration"
      "operationId": "post__api_v1_house-registrations_id_cancel"
      "description": "Required role: Authenticated user, including without membership. gRPC: Identity.HouseService.CancelHouseRegistration.  List responses are bounded to 100 records (members: 500). Mutation requests require a JSON object, including {} for commands without fields."
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/HouseRegistration"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "Origin"
          "in": "header"
          "required": true
          "schema":
            "type": "string"
            "format": "uri"
          "description": "Exact origin from TRUSTED_ORIGINS. Missing or foreign Origin is rejected (CSRF protection)."
        -
          "name": "Idempotency-Key"
          "in": "header"
          "schema":
            "type": "string"
            "format": "uuid"
          "description": "Optional. Scope: user + active house. Key binds method, path and compact JSON bytes (key order significant). Same request replays success for 24h; changed payload 409 IDEMPOTENCY_KEY_REUSED; concurrent/ambiguous outcome 409 IDEMPOTENCY_IN_PROGRESS. Reconcile before creating a new key; no exactly-once guarantee."
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "type": "object"
              "additionalProperties": false
              "properties": {}
              "required": []
...
"components":
  "securitySchemes":
    "SessionAuth":
      "type": "apiKey"
      "in": "cookie"
      "name": "sq_session"
    "MaxWebhookSecret":
      "type": "apiKey"
      "in": "header"
      "name": "X-Max-Bot-Api-Secret"
  "schemas":
...
    "HouseRegistration":
      "type": "object"
      "additionalProperties": false
      "properties":
        "id":
          "type": "string"
          "format": "uuid"
        "applicant_user_id":
          "type": "string"
          "format": "uuid"
        "requested_name":
          "type": "string"
        "original_address":
          "type": "string"
        "city":
          "type": "string"
        "status":
          "$ref": "#/components/schemas/HouseRegistrationStatus"
        "resulting_house_id":
          "type": "string"
          "description": "Created house UUID; empty before approval"
        "rejection_reason":
          "type": "string"
        "created_at":
          "type":
            - "string"
            - "null"
          "format": "date-time"
        "reviewed_at":
          "type":
            - "string"
            - "null"
          "format": "date-time"
        "applicant_display_name":
          "type": "string"
      "required":
        - "id"
        - "applicant_user_id"
        - "requested_name"
        - "original_address"
        - "city"
        - "status"
        - "resulting_house_id"
        - "rejection_reason"
        - "applicant_display_name"
...
  "responses":
    "Error400":
      "description": "Invalid JSON, enum, identifier or pagination"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error401":
      "description": "Missing/invalid/expired session or initData"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error403":
      "description": "Forbidden membership, role or Origin"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error404":
      "description": "Resource not found"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error409":
      "description": "Conflict, invalid transition or idempotency reservation"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error413":
      "description": "Body exceeds 256 KiB"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error415":
      "description": "application/json required"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error429":
      "description": "Rate limit exceeded"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error500":
      "description": "Internal error"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error502":
      "description": "Invalid upstream response"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
...