# services/max-gateway/internal/transport/http/house_generated.go
// Code generated by scripts/generate-house-http.py; DO NOT EDIT.
package http

import (
	"encoding/json"
	ipb "github.com/ZheglY/SmartQuarter/services/max-gateway/internal/gen/smartquarter/identity/v1"
	"github.com/ZheglY/SmartQuarter/services/max-gateway/internal/identity"
	"google.golang.org/protobuf/encoding/protojson"
	"google.golang.org/protobuf/proto"
	"net/http"
	"net/url"
)

func (a *API) houseResponse(w http.ResponseWriter, r *http.Request, code int, v proto.Message, e error) {
	if e != nil {
		a.rpcError(w, r, e)
		return
	}
	raw, e := (protojson.MarshalOptions{UseProtoNames: true, EmitDefaultValues: true}).Marshal(v)
	if e != nil {
		a.fail(w, r, 502, "INVALID_UPSTREAM_RESPONSE", "invalid Identity response")
		return
	}
	write(w, code, json.RawMessage(raw))
}
func (a *API) houseRoutes(mux *http.ServeMux) {
	mux.HandleFunc("POST /api/v1/house-registrations", a.origin(a.authenticate(false, false, a.idempotent(a.houseCreateHouseRegistration))))
	mux.HandleFunc("GET /api/v1/house-registrations/{id}", a.origin(a.authenticate(false, false, a.houseGetHouseRegistration)))
	mux.HandleFunc("GET /api/v1/house-registrations", a.origin(a.authenticate(false, false, a.houseListMyHouseRegistrations)))
	mux.HandleFunc("POST /api/v1/house-registrations/{id}/cancel", a.origin(a.authenticate(false, false, a.idempotent(a.houseCancelHouseRegistration))))
	mux.HandleFunc("GET /api/v1/admin/house-registrations", a.origin(a.authenticate(false, false, a.housePermission("admin", a.houseListPendingHouseRegistrations))))
	mux.HandleFunc("POST /api/v1/admin/house-registrations/{id}/approve", a.origin(a.authenticate(false, false, a.housePermission("admin", a.idempotent(a.houseApproveHouseRegistration)))))
	mux.HandleFunc("POST /api/v1/admin/house-registrations/{id}/reject", a.origin(a.authenticate(false, false, a.housePermission("admin", a.idempotent(a.houseRejectHouseRegistration)))))
	mux.HandleFunc("GET /api/v1/houses/search", a.origin(a.authenticate(false, false, a.houseSearchHouses)))
	mux.HandleFunc("POST /api/v1/houses/{house_id}/join-requests", a.origin(a.authenticate(false, false, a.idempotent(a.houseCreateJoinRequest))))
	mux.HandleFunc("GET /api/v1/join-requests", a.origin(a.authenticate(false, false, a.houseListMyJoinRequests)))
	mux.HandleFunc("POST /api/v1/join-requests/{id}/cancel", a.origin(a.authenticate(false, false, a.idempotent(a.houseCancelJoinRequest))))
	mux.HandleFunc("GET /api/v1/chairman/join-requests", a.origin(a.authenticate(false, false, a.housePermission("manager", a.houseListHouseJoinRequests))))
	mux.HandleFunc("POST /api/v1/chairman/join-requests/{id}/approve", a.origin(a.authenticate(false, false, a.housePermission("manager", a.idempotent(a.houseApproveJoinRequest)))))
	mux.HandleFunc("POST /api/v1/chairman/join-requests/{id}/reject", a.origin(a.authenticate(false, false, a.housePermission("manager", a.idempotent(a.houseRejectJoinRequest)))))
	mux.HandleFunc("POST /api/v1/chairman/invitations", a.origin(a.authenticate(false, false, a.housePermission("manager", a.idempotent(a.houseCreateHouseInvitation)))))
	mux.HandleFunc("GET /api/v1/chairman/invitations", a.origin(a.authenticate(false, false, a.housePermission("manager", a.houseListHouseInvitations))))
	mux.HandleFunc("GET /api/v1/invitations/{id}", a.origin(a.authenticate(false, false, a.housePermission("manager", a.houseGetHouseInvitation))))
	mux.HandleFunc("POST /api/v1/chairman/invitations/{id}/revoke", a.origin(a.authenticate(false, false, a.housePermission("manager", a.idempotent(a.houseRevokeHouseInvitation)))))
	mux.HandleFunc("POST /api/v1/invitations/redeem", a.origin(a.authenticate(false, false, a.idempotent(a.houseRedeemHouseInvitation))))
	mux.HandleFunc("POST /api/v1/invitations/preview", a.origin(a.authenticate(false, false, a.idempotent(a.housePreviewHouseInvitation))))
	mux.HandleFunc("POST /api/v1/chairman/transfers", a.origin(a.authenticate(false, false, a.housePermission("manager", a.idempotent(a.houseCreateChairmanTransfer)))))
	mux.HandleFunc("GET /api/v1/chairman/transfers", a.origin(a.authenticate(false, false, a.houseListChairmanTransfers)))
	mux.HandleFunc("POST /api/v1/chairman/transfers/{id}/accept", a.origin(a.authenticate(false, false, a.idempotent(a.houseAcceptChairmanTransfer))))
	mux.HandleFunc("POST /api/v1/chairman/transfers/{id}/reject", a.origin(a.authenticate(false, false, a.idempotent(a.houseRejectChairmanTransfer))))
	mux.HandleFunc("POST /api/v1/chairman/transfers/{id}/cancel", a.origin(a.authenticate(false, false, a.idempotent(a.houseCancelChairmanTransfer))))
	mux.HandleFunc("GET /api/v1/chairman/members", a.origin(a.authenticate(false, false, a.housePermission("manager", a.houseListHouseMembers))))
	mux.HandleFunc("POST /api/v1/chairman/members/{id}/deactivate", a.origin(a.authenticate(false, false, a.housePermission("manager", a.idempotent(a.houseDeactivateMembership)))))
	mux.HandleFunc("POST /api/v1/chairman/members/{id}/reactivate", a.origin(a.authenticate(false, false, a.housePermission("manager", a.idempotent(a.houseReactivateMembership)))))
	mux.HandleFunc("DELETE /api/v1/chairman/members/{id}", a.origin(a.authenticate(false, false, a.housePermission("manager", a.idempotent(a.houseRemoveMembership)))))
	mux.HandleFunc("GET /api/v1/house-access", a.origin(a.authenticate(false, false, a.houseGetHouseAccessState)))
	mux.HandleFunc("GET /api/v1/notifications/settings", a.origin(a.authenticate(false, false, a.houseGetNotificationPreferences)))
	mux.HandleFunc("PUT /api/v1/notifications/settings", a.origin(a.authenticate(false, false, a.idempotent(a.houseUpdateNotificationPreferences))))
}
func (a *API) houseCreateHouseRegistration(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	var body struct {
		Name    string `json:"name"`
		Address string `json:"address"`
		City    string `json:"city"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.CreateHouseRegistrationRequest{Name: body.Name, Address: body.Address, City: body.City}
	v, e := a.House.CreateHouseRegistration(r.Context(), req)
	a.houseResponse(w, r, 202, v, e)
}
func (a *API) houseGetHouseRegistration(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	req := &ipb.GetHouseRegistrationRequest{Id: r.PathValue("id")}
	v, e := a.House.GetHouseRegistration(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseListMyHouseRegistrations(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	req := &ipb.ListMyHouseRegistrationsRequest{}
	v, e := a.House.ListMyHouseRegistrations(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseCancelHouseRegistration(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct{}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.CancelHouseRegistrationRequest{Id: r.PathValue("id")}
	v, e := a.House.CancelHouseRegistration(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseListPendingHouseRegistrations(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	req := &ipb.ListPendingHouseRegistrationsRequest{}
	v, e := a.House.ListPendingHouseRegistrations(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseApproveHouseRegistration(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct {
		Reason string `json:"reason"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.ApproveHouseRegistrationRequest{Id: r.PathValue("id"), Reason: body.Reason}
	v, e := a.House.ApproveHouseRegistration(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseRejectHouseRegistration(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct {
		Reason string `json:"reason"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.RejectHouseRegistrationRequest{Id: r.PathValue("id"), Reason: body.Reason}
	v, e := a.House.RejectHouseRegistration(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseSearchHouses(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	req := &ipb.SearchHousesRequest{Query: r.URL.Query().Get("query")}
	v, e := a.House.SearchHouses(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseCreateJoinRequest(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("house_id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct{}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.CreateJoinRequestRequest{HouseId: r.PathValue("house_id")}
	v, e := a.House.CreateJoinRequest(r.Context(), req)
	a.houseResponse(w, r, 202, v, e)
}
func (a *API) houseListMyJoinRequests(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	req := &ipb.ListMyJoinRequestsRequest{}
	v, e := a.House.ListMyJoinRequests(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseCancelJoinRequest(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct{}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.CancelJoinRequestRequest{Id: r.PathValue("id")}
	v, e := a.House.CancelJoinRequest(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseListHouseJoinRequests(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	req := &ipb.ListHouseJoinRequestsRequest{}
	v, e := a.House.ListHouseJoinRequests(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseApproveJoinRequest(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct {
		Reason string `json:"reason"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.ApproveJoinRequestRequest{Id: r.PathValue("id"), Reason: body.Reason}
	v, e := a.House.ApproveJoinRequest(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseRejectJoinRequest(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct {
		Reason string `json:"reason"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.RejectJoinRequestRequest{Id: r.PathValue("id"), Reason: body.Reason}
	v, e := a.House.RejectJoinRequest(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseCreateHouseInvitation(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	var body struct {
		ExpiresInHours int32 `json:"expires_in_hours"`
		MaxUses        int32 `json:"max_uses"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.CreateHouseInvitationRequest{ExpiresInHours: body.ExpiresInHours, MaxUses: body.MaxUses}
	v, e := a.House.CreateHouseInvitation(r.Context(), req)
	if e == nil && a.Config.BotUsername != "" {
		v.DeepLink = "https://max.ru/" + url.PathEscape(a.Config.BotUsername) + "?startapp=invite_" + v.Token
	}
	a.houseResponse(w, r, 201, v, e)
}
func (a *API) houseListHouseInvitations(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	req := &ipb.ListHouseInvitationsRequest{}
	v, e := a.House.ListHouseInvitations(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseGetHouseInvitation(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	req := &ipb.GetHouseInvitationRequest{Id: r.PathValue("id")}
	v, e := a.House.GetHouseInvitation(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseRevokeHouseInvitation(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct{}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.RevokeHouseInvitationRequest{Id: r.PathValue("id")}
	v, e := a.House.RevokeHouseInvitation(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseRedeemHouseInvitation(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	var body struct {
		Token string `json:"token"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.RedeemHouseInvitationRequest{Token: body.Token}
	v, e := a.House.RedeemHouseInvitation(r.Context(), req)
	a.houseResponse(w, r, 202, v, e)
}
func (a *API) housePreviewHouseInvitation(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	var body struct {
		Token string `json:"token"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.PreviewHouseInvitationRequest{Token: body.Token}
	v, e := a.House.PreviewHouseInvitation(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseCreateChairmanTransfer(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	var body struct {
		TargetUserId string `json:"target_user_id"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.CreateChairmanTransferRequest{TargetUserId: body.TargetUserId}
	v, e := a.House.CreateChairmanTransfer(r.Context(), req)
	a.houseResponse(w, r, 202, v, e)
}
func (a *API) houseListChairmanTransfers(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	req := &ipb.ListChairmanTransfersRequest{}
	v, e := a.House.ListChairmanTransfers(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseAcceptChairmanTransfer(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct{}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.AcceptChairmanTransferRequest{Id: r.PathValue("id")}
	v, e := a.House.AcceptChairmanTransfer(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseRejectChairmanTransfer(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct{}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.RejectChairmanTransferRequest{Id: r.PathValue("id")}
	v, e := a.House.RejectChairmanTransfer(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseCancelChairmanTransfer(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct{}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.CancelChairmanTransferRequest{Id: r.PathValue("id")}
	v, e := a.House.CancelChairmanTransfer(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseListHouseMembers(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	req := &ipb.ListHouseMembersRequest{}
	v, e := a.House.ListHouseMembers(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseDeactivateMembership(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct {
		PlatformAdminOverride bool `json:"platform_admin_override"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.DeactivateMembershipRequest{Id: r.PathValue("id"), PlatformAdminOverride: body.PlatformAdminOverride}
	v, e := a.House.DeactivateMembership(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseReactivateMembership(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct {
		PlatformAdminOverride bool `json:"platform_admin_override"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.ReactivateMembershipRequest{Id: r.PathValue("id"), PlatformAdminOverride: body.PlatformAdminOverride}
	v, e := a.House.ReactivateMembership(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseRemoveMembership(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	if !identity.ValidID(r.PathValue("id")) {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "invalid resource id")
		return
	}
	var body struct {
		PlatformAdminOverride bool `json:"platform_admin_override"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	req := &ipb.RemoveMembershipRequest{Id: r.PathValue("id"), PlatformAdminOverride: body.PlatformAdminOverride}
	v, e := a.House.RemoveMembership(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseGetHouseAccessState(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	req := &ipb.GetHouseAccessStateRequest{}
	v, e := a.House.GetHouseAccessState(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseGetNotificationPreferences(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	req := &ipb.GetNotificationPreferencesRequest{}
	v, e := a.House.GetNotificationPreferences(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}
func (a *API) houseUpdateNotificationPreferences(w http.ResponseWriter, r *http.Request) {
	if a.House == nil {
		a.fail(w, r, 503, "DEPENDENCY_UNAVAILABLE", "house workflow unavailable")
		return
	}
	var body struct {
		NotificationsEnabled             *bool `json:"notifications_enabled"`
		IssueNotificationsEnabled        *bool `json:"issue_notifications_enabled"`
		AnnouncementNotificationsEnabled *bool `json:"announcement_notifications_enabled"`
		MembershipNotificationsEnabled   *bool `json:"membership_notifications_enabled"`
		BotNotificationsEnabled          *bool `json:"bot_notifications_enabled"`
	}
	if !a.decode(w, r, &body) {
		return
	}
	if body.NotificationsEnabled == nil {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "all preferences required")
		return
	}
	if body.IssueNotificationsEnabled == nil {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "all preferences required")
		return
	}
	if body.AnnouncementNotificationsEnabled == nil {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "all preferences required")
		return
	}
	if body.MembershipNotificationsEnabled == nil {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "all preferences required")
		return
	}
	if body.BotNotificationsEnabled == nil {
		a.fail(w, r, 400, "INVALID_ARGUMENT", "all preferences required")
		return
	}
	req := &ipb.UpdateNotificationPreferencesRequest{NotificationsEnabled: *body.NotificationsEnabled, IssueNotificationsEnabled: *body.IssueNotificationsEnabled, AnnouncementNotificationsEnabled: *body.AnnouncementNotificationsEnabled, MembershipNotificationsEnabled: *body.MembershipNotificationsEnabled, BotNotificationsEnabled: *body.BotNotificationsEnabled}
	v, e := a.House.UpdateNotificationPreferences(r.Context(), req)
	a.houseResponse(w, r, 200, v, e)
}